+++ src/euphorie/client/tests/test_survey.py
@@ -180,8 +180,6 @@ class completion_percentage_tests(EuphorieIntegrationTestCase):
         self.session.flush()
 
     def test_no_risks(self):
-        with self._get_view("webhelpers", self.portal.client) as view:
-            view.update_completion_percentage(self.survey)
         self.assertEqual(self.survey.completion_percentage, 0)
 
     def test_half_complete(self):
@@ -199,8 +197,6 @@ class completion_percentage_tests(EuphorieIntegrationTestCase):
         self.q1.postponed = False
         self.q1.identification = "no"
 
-        with self._get_view("webhelpers", self.portal.client) as view:
-            view.update_completion_percentage(self.survey)
         self.assertEqual(self.survey.completion_percentage, 50)
 
     def test_two_thirds_complete(self):
@@ -226,8 +222,6 @@ class completion_percentage_tests(EuphorieIntegrationTestCase):
         self.q21.postponed = False
         self.q21.identification = "yes"
 
-        with self._get_view("webhelpers", self.portal.client) as view:
-            view.update_completion_percentage(self.survey)
         self.assertEqual(self.survey.completion_percentage, 67)
 
     def test_optional_module(self):
@@ -247,8 +241,6 @@ class completion_percentage_tests(EuphorieIntegrationTestCase):
         self.q1.identification = "yes"
         self.mod2.skip_children = True
 
-        with self._get_view("webhelpers", self.portal.client) as view:
-            view.update_completion_percentage(self.survey)
         self.assertEqual(self.survey.completion_percentage, 100)
 
     def test_optional_submodule(self):
@@ -271,8 +263,6 @@ class completion_percentage_tests(EuphorieIntegrationTestCase):
         self.q1.identification = "yes"
         self.mod21.skip_children = True
 
-        with self._get_view("webhelpers", self.portal.client) as view:
-            view.update_completion_percentage(self.survey)
         self.assertEqual(self.survey.completion_percentage, 100)
 
     def test_optional_module_with_submodule(self):
@@ -296,6 +286,4 @@ class completion_percentage_tests(EuphorieIntegrationTestCase):
         self.q1.identification = "yes"
         self.mod2.skip_children = True
         self.mod21.skip_children = False
-        with self._get_view("webhelpers", self.portal.client) as view:
-            view.update_completion_percentage(self.survey)
         self.assertEqual(self.survey.completion_percentage, 100)