+++ docs/changes.rst
@@ -4,6 +4,8 @@ Changelog
 14.1.5 (unreleased)
 -------------------
 
+- Deprecate the @@update-completion-percentage view
+  `#419 <https://github.com/euphorie/Euphorie/pull/419>`_.
 - “Show more” button on tool view if there are existing sessions.
 
 
@@ -1744,6 +1746,7 @@ Bugfixes
 
 6.0rc3 - April 23, 2013
 -----------------------
+
 
 - Update Dutch, Latvian, Lithuanian and Finnish translations.
 - Use https in stylesheets (for google fonts).